+++ python/src/client.py
@@ -9,6 +9,7 @@ from system import get_system_id
 class Client:
     def __init__(self, websocket):
         self.cache = {}
+        self.requests = {}
         self.system_id = get_system_id()
         self.transaction = 1
         self.websocket = websocket
@@ -28,11 +29,13 @@ class Client:
 
     def add_context(self, node_id, id, name):
         args = [self.system_id, node_id, id, name]
-        return self.send('json', 'contexts', args)
+        transaction = self.send('json', 'contexts', args)
+        self.wait_for_response(transaction)
 
     def add_node(self, id, name, upstream=False, token=None):
         args = [self.system_id, id, name, upstream]
-        return self.send('json', 'nodes', args)
+        transaction = self.send('json', 'nodes', args)
+        self.wait_for_response(transaction)
 
     def add_system(self, id=None, name=None):
         if id is None:
@@ -61,7 +64,9 @@ class Client:
             'format': format,
             'command': cmd,
         })
-        return self.websocket.send(message)
+        self.websocket.send(message)
+        self.requests[self.transaction] = None
+        return self.transaction
 
     def stats(self):
         return self.cache['stats']
@@ -75,9 +80,29 @@ class Client:
             if self.websocket.state == State.OPEN:
                 return
             time.sleep(0.1)
-        raise Exception('Timed out waiting for open')
+
+    def wait_for_response(self, transaction, timeout_secs = 5):
+        start_time = time.time()
+        while time.time() - start_time < timeout_secs:
+            if transaction not in self.requests:
+                raise Exception('No such transaction')
+            response = self.requests[transaction]
+            if response is not None:
+                if 'error' in response:
+                    error = response['error']
+                    if error is not None:
+                        raise Exception(error)
+                return
+            time.sleep(0.1)
+        raise Exception('Timed out waiting for response')
 
 def receive_messages(self):
     for message in self.websocket:
         data = json.loads(message)
+        if 'transaction' in data:
+            transaction = int(data['transaction'])
+            if 'response' in data:
+                response = data['response']
+                if response == '':
+                    self.requests[transaction] = {'error': None}
         self.cache.update(data)